pyproject.toml
[project]
name = "image-manipulation-datasets"
version = "1.0.1"
description = "A collection of image manipulation dataset classes implemented in PyTorch"
authors = [{ name = "Spencer Cain", email = "cainspencerm@protonmail.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch (>=1.11.0,<2.0.0)",
    "pillow (>=9.1.1,<10.0.0)",
    "numpy (>=1.22.4,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "image-manipulation-datasets"
version = "1.0.1"
packages = [{ include = "imds", from = "src" }]

[tool.poetry.group.dev.dependencies]
mypy = "^1.16.1"
types-pillow = "^10.2.0.20240822"

[tool.semantic_release]
version_toml = [
    "pyproject.toml:project.version",
    "pyproject.toml:tool.poetry.version",
]
